Add 'balloon' target and unfinished 'lora' driver
This commit adds a 'balloon' target for fmu-v5. This target was prepared
for reporting telemetry from a stratospheric balloon flight.

Besides the balloon target, also included is an unfinished driver for
sending messages as end-device in a LoRaWAN network. This involves
talking to an SX1262 chip appropriately wired to the SPI5 port. To free
one wire on the SPI connector, one of chip select signals is disabled
in the fmu-v5 board config.

The 'lora' driver contains a copy of Semtech's LoRaWAN end-device
library to which unknown changes were made.